feat: Add createdBy authorization to EditServiceRequest page
This commit adds authorization to the EditServiceRequest page in the ServiceRequestResource. The page now checks if the current user is the creator of the service request before allowing access. If the user is not the creator, a 403 error is returned. This change enhances the security of the application by restricting unauthorized access to edit service requests.